// web-application/src/types-openapi/api/models/VoucherResponse.ts
export interface VoucherResponse {
    /**
     * 
     * @type {number}
     * @memberof VoucherResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof VoucherResponse
     */
    code?: string | null;
    /**
     * 
     * @type {string}
     * @memberof VoucherResponse
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof VoucherResponse
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof VoucherResponse
     */
    discountType?: string | null;
    /**
     * 
     * @type {number}
     * @memberof VoucherResponse
     */
    discountValue?: number;
    /**
     * 
     * @type {number}
     * @memberof VoucherResponse
     */
    discountPercentage?: number | null;
    /**
     * 
     * @type {number}
     * @memberof VoucherResponse
     */
    maxDiscountValue?: number | null;
    /**
     * 
     * @type {number}
     * @memberof VoucherResponse
     */
    minOrderValue?: number | null;
    /**
     * 
     * @type {Date}
     * @memberof VoucherResponse
     */
    startAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof VoucherResponse
     */
    endAt?: Date;
    /**
     * 
     * @type {number}
     * @memberof VoucherResponse
     */
    usageLimitTotal?: number;
    /**
     * 
     * @type {number}
     * @memberof VoucherResponse
     */
    usageLimitPerUser?: number;
    /**
     * 
     * @type {number}
     * @memberof VoucherResponse
     */
    usedCount?: number;
    /**
     * 
     * @type {boolean}
     * @memberof VoucherResponse
     */
    isActive?: boolean;
    /**
     * 
     * @type {Array<VoucherTimeRuleDto>}
     * @memberof VoucherResponse
     */
    timeRules?: Array<VoucherTimeRuleDto> | null;
    /**
     * 
     * @type {Array<VoucherUserRuleDto>}
     * @memberof VoucherResponse
     */
    userRules?: Array<VoucherUserRuleDto> | null;
}